Move publicly accessed objects to public namespaces.
