+++ src/EFCore.MySql/Storage/ServerVersion.Support.cs
@@ -1,7 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
+using Pomelo.EntityFrameworkCore.MySql.Storage.Internal;
 
-namespace Pomelo.EntityFrameworkCore.MySql.Storage.Internal
+namespace Pomelo.EntityFrameworkCore.MySql.Storage
 {
     public partial class ServerVersion
     {
